+++ bobtemplates/hooks.py
@@ -2,7 +2,9 @@
 # -*- coding: utf-8 -*-
 """Render bobtemplates.plone hooks.
 """
+from mrbob.bobexceptions import SkipQuestion
 from mrbob.bobexceptions import ValidationError
+from mrbob.hooks import validate_choices
 
 import os
 import shutil
@@ -101,6 +103,21 @@ def post_plone_version(configurator, question, answer):
     return answer
 
 
+def post_type(configurator, question, answer):
+    """Skip questions depending on the type answer.
+    """
+    value = validate_choices(configurator, question, answer)
+    if value != u'Dexterity':
+        configurator.variables['package.dexterity_type_name'] = ''
+        configurator.variables['package.dexterity_type_name_lower'] = ''
+    return value
+
+
+def pre_dexterity_type_name(configurator, question):
+    if configurator.variables['package.type'] != 'Dexterity':
+        raise SkipQuestion
+
+
 def prepare_render(configurator):
     """Some variables to make templating easier.
 
@@ -162,6 +179,17 @@ def prepare_render(configurator):
             configurator.variables['package.namespace'])
     configurator.variables['package.namespace_packages'] = namespace_packages
 
+    if configurator.variables.get('package.dexterity_type_name'):
+        configurator.variables[
+            'package.dexterity_type_name_lower'
+        ] = configurator.variables['package.dexterity_type_name'].lower()
+    else:
+        # We have to make sure those variables are always set because we are
+        # going to create files that contain those variables. Even if we
+        # remove those files afterwards. This is just how mr.bob rolls.
+        configurator.variables['package.dexterity_type_name'] = ''
+        configurator.variables['package.dexterity_type_name_lower'] = ''
+
 
 def cleanup_package(configurator):
     """Cleanup and make nested if needed.
@@ -218,16 +246,18 @@ def cleanup_package(configurator):
     # find out what to delete
     to_delete = []
 
-    if not configurator.variables['package.example']:
-        to_delete.extend([
-            "{0}/browser/templates",
-            "{0}/browser/views.py",
-        ])
-
-    if not configurator.variables['package.theme']:
+    if configurator.variables['package.type'] != u'Theme':
         to_delete.extend([
             "{0}/theme",
             "{0}/profiles/default/theme.xml",
+        ])
+
+    if configurator.variables['package.type'] != u'Dexterity':
+        to_delete.extend([
+            "{0}/profiles/default/types.xml",
+            "{0}/profiles/default/types",
+            "{0}/tests/test_.py",
+            "{0}/tests/robot/test_.robot",
         ])
 
     # remove parts